requirements.txt
asgiref==3.8.1
asttokens==3.0.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Django==5.0.6
django-admin-soft-dashboard==1.0.18
django-cors-headers==4.3.1
django-filter==24.2
django-jazzmin==3.0.1
django-phonenumber-field==7.3.0
django-semantic-admin==0.6.2
django-semantic-forms==0.1.8
django-taggit==6.1.0
djangorestframework==3.15.1
executing==2.1.0
fonttools==4.55.0
icecream==2.1.3
idna==3.7
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.3
packaging==24.2
phonenumbers==8.13.36
pillow==10.3.0
psycopg2==2.9.9
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
toposort==1.10
tzdata==2024.1
urllib3==2.2.1
